requirements.txt
flask~=1.1.2
celery~=5.0.5
pyyaml~=5.4.1
redis~=3.5.3